My experiment with "cvEqualizeHist()" worked out pretty well for the mouth in "detectMouthFeaturePoints()". In "data/image_0316.jpg" the mouth wasn't detected at first, but with the histogram equalization, the mouth contour is much better. I also applied the "cvEqualizeHist()" function to "detectEyeFeaturePoints()". Evalutaion is still needed.
What's still missing from the mouth detection is the flood fill operation as described on page 662! If we don't implement it and fine tune the mouth algorithm, we will only detect the inner shape of a mouth. At least that's what I experienced when capturing with my webcam.

The data/image_0316.jpg was a pretty good example for a requirement of fine-tuning our "detectEyeFeaturePoints()" function. Before this commit. The function only detected the eyebrows instead of the eyes. By ensuring the eye center point lies within the rectangle which surrounds the contour, I ensured that the eye is detected in favor of the eyebrow.

We should consider the processing speed and evaluate if our custom "stretchContrast()" function is obsolete because of "cvEqualizeHist()" now. The OpenCV function seems to do a much better job than our implementation even though they're both in use at the moment.

Buenas noches!