Able to pass custom struct to DLL and vice versa (without ugly unsafe code!!!). Some update on ANN drawing.
The "FreeCapture" Method from KluSharp/Klu.cs is being misused at the moment to pass a custom two-element struct to the DLL to calculate the sum of the two elements. The result is printed to console.

There is still a slight little gap between the ANN canvas and it's border. I couldn't make it disappear. Ask Marc or Chris for a fix! Still the layout is not fixed. There are missing labels and bindings to ann.NumLayers and ann.NumNeurons. Again, ask Marc for this!